+++ app/android/src/main/java/app/campfire/android/di/ActivityComponent.kt
@@ -3,8 +3,8 @@ package app.campfire.android.di
 import android.app.Activity
 import androidx.core.os.ConfigurationCompat
 import app.campfire.account.api.UserSessionManager
+import app.campfire.audioplayer.cast.CastController
 import app.campfire.audioplayer.impl.MediaControllerConnector
-import app.campfire.audioplayer.impl.cast.MediaRouterCastController
 import app.campfire.audioplayer.offline.OfflineDownloadManager
 import app.campfire.common.root.CampfireContent
 import app.campfire.core.ComponentActivityPlugin
@@ -24,7 +24,7 @@ import me.tatarka.inject.annotations.Provides
 interface ActivityComponent {
   val campfireContent: CampfireContent
   val mediaControllerConnector: MediaControllerConnector
-  val mediaRouterCastController: MediaRouterCastController
+  val castController: CastController
   val componentActivityPlugins: Set<ComponentActivityPlugin>
   val offlineDownloadManager: OfflineDownloadManager
   val userSessionManager: UserSessionManager